feat: adds toggleMounting feature
