+++ demo/index.html
@@ -24,6 +24,7 @@
 		
 			<button class="toggle__show">Show</button>
 			<button class="toggle__hide">Hide</button>
+			<button class="toggle__state">Toggle</button>
 			<div class="toggle__element" data-transition="fade">
 				<h2>This is the toggle element</h2>
 				<p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. Delectus, maiores at! Quia illum ipsum, minus ratione totam reprehenderit. Totam, dolorum. Dicta dolores sint totam animi non a labore accusamus hic.</p>
