+++ demo/scripts/toggle.js
@@ -1,7 +1,14 @@
-import { mount, unmount, Watcher, animateHeightAuto } from '../../lib';
+import {
+  mount,
+  unmount,
+  toggleMounting,
+  Watcher,
+  animateHeightAuto,
+} from '../../lib';
 
 const toggleShow = new Watcher('.toggle__show');
 const toggleHide = new Watcher('.toggle__hide');
+const toggleState = new Watcher('.toggle__state');
 const toggleElement = document.querySelector('.toggle__element');
 
 toggleShow.click(() => {
@@ -13,5 +20,11 @@ toggleShow.click(() => {
 toggleHide.click(() => {
   animateHeightAuto('.toggle', () => {
     unmount(toggleElement);
+  });
+});
+
+toggleState.click(() => {
+  animateHeightAuto('.toggle', () => {
+    toggleMounting(toggleElement);
   });
 });
